swap/requirements.txt
python-dotenv==1.0.0
requests==2.32.3
gspread==6.1.4
aiogram==2.25.1
schedule==1.2.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
httplib2==0.22.0